Remove tests for deprecated Grain objects
